feat: split server vs client components render
