+++ src/app/signin/page.tsx
@@ -1,38 +1,7 @@
-"use client";
-
-import Image from "next/image";
-import { toast } from "sonner";
 import { ToolHiveIcon } from "@/components/icons";
-import { Button } from "@/components/ui/button";
-import { authClient } from "@/lib/auth/auth-client";
-import { OIDC_PROVIDER_ID } from "@/lib/auth/constants";
+import { SignInButton } from "./signin-button";
 
 export default function SignInPage() {
-  const handleOIDCSignIn = async () => {
-    try {
-      const { error } = await authClient.signIn.oauth2({
-        providerId: OIDC_PROVIDER_ID,
-        callbackURL: "/catalog",
-      });
-
-      if (error) {
-        toast.error("Signin failed", {
-          description:
-            error.message ||
-            "An error occurred during signin. Please try again.",
-        });
-        return;
-      }
-    } catch (error) {
-      const errorMessage =
-        error instanceof Error ? error.message : "An unexpected error occurred";
-
-      toast.error("Signin error", {
-        description: errorMessage,
-      });
-    }
-  };
-
   return (
     <div className="flex h-screen w-full">
       <div className="hidden md:flex w-1/2 bg-muted/80 border-r border-border items-start p-10">
@@ -51,20 +20,7 @@ export default function SignInPage() {
             </p>
           </div>
 
-          <Button
-            onClick={handleOIDCSignIn}
-            className="w-full h-9 gap-2"
-            size="default"
-          >
-            <Image
-              src="/okta-icon.svg"
-              alt="Okta"
-              width={16}
-              height={16}
-              className="shrink-0"
-            />
-            <span>Okta</span>
-          </Button>
+          <SignInButton />
         </div>
       </div>
     </div>
